data model refactor

+++ src/services/databaseService.ts
@@ -15,30 +15,53 @@ import { db } from "../firebase/config";
 
 import type { Expense, Payment, Participant, SharedExpense } from "../types";
 
+const USERS_COLLECTION_NAME = "users";
+
+const BASE_COLLECTION_PATH = `environments/${
+  import.meta.env.VITE_FIRESTORE_DATA_ID
+}/${USERS_COLLECTION_NAME}`;
+
+const CONTACTS_COLLECTION_NAME = "contacts";
+const EXPENSES_COLLECTION_NAME = "expenses";
+const PAYMENTS_COLLECTION_NAME = "payments";
 const SHARED_EXPENSES_COLLECTION_NAME = "sharedExpenses";
 
-// Participant Operations
+// Contacts Operations
+function getContactsPath(uid: string) {
+  return `${BASE_COLLECTION_PATH}/${uid}/${CONTACTS_COLLECTION_NAME}`;
+}
+function getContactsRef(collectionPath: string) {
+  return collection(db, collectionPath);
+}
+
 export const participantService = {
   async createParticipant(
-    participant: Omit<Participant, "id">
+    participant: Omit<Participant, "id">,
+    uid: string
   ): Promise<string> {
-    const docRef = await addDoc(collection(db, "participants"), {
+    const contactsPath = getContactsPath(uid);
+    const contactsRef = getContactsRef(contactsPath);
+
+    const docRef = await addDoc(contactsRef, {
       ...participant,
       createdAt: Timestamp.now(),
     });
+
     return docRef.id;
   },
 
-  async createParticipantList(): Promise<string[]> {
+  async createParticipantList(uid: string): Promise<string[]> {
     const participantList = [
-      { name: "Fer" },
-      { name: "Seba" },
-      { name: "Nata" },
+      { name: "Seba", email: "seba@example.com" },
+      { name: "Nata", email: "nata@example.com" },
     ];
+
+    const contactsPath = getContactsPath(uid);
+    const contactsRef = getContactsRef(contactsPath);
 
     const promises = participantList.map(
       (participant: Omit<Participant, "id">) =>
-        addDoc(collection(db, "participants"), {
+        addDoc(contactsRef, {
           ...participant,
           createdAt: Timestamp.now(),
         })
@@ -49,8 +72,10 @@ export const participantService = {
     return docRef.map((ref) => ref.id);
   },
 
-  async getParticipants(): Promise<Participant[]> {
-    const querySnapshot = await getDocs(query(collection(db, "participants")));
+  async getParticipants(uid: string): Promise<Participant[]> {
+    const contactsPath = getContactsPath(uid);
+    const contactsRef = getContactsRef(contactsPath);
+    const querySnapshot = await getDocs(query(contactsRef));
 
     return querySnapshot.docs.map(
       (doc) =>
@@ -65,23 +90,23 @@ export const participantService = {
 // --------------------------------------------------
 
 // Expense Operations
-const EXPENSES_COLLECTION_NAME = "expenses";
 
-function getExpensesCollectionPath(sharedExpenseId: string) {
-  return `environments/${
-    import.meta.env.VITE_FIRESTORE_DATA_ID
-  }/${SHARED_EXPENSES_COLLECTION_NAME}/${sharedExpenseId}/${EXPENSES_COLLECTION_NAME}`;
+function getExpensesPath(sharedExpenseId: string, uid: string) {
+  return `${BASE_COLLECTION_PATH}/${uid}/${SHARED_EXPENSES_COLLECTION_NAME}/${sharedExpenseId}/${EXPENSES_COLLECTION_NAME}`;
 }
 
-async function getExpensesCollectionRef(collectionPath: string) {
+async function getExpensesRef(collectionPath: string) {
   return collection(db, collectionPath);
 }
 
 export const expenseService = {
   // Create new expense
-  async createExpense(expense: Omit<Expense, "id">): Promise<string> {
-    const collectionPath = getExpensesCollectionPath(expense.sharedExpenseId);
-    const collectionRef = await getExpensesCollectionRef(collectionPath);
+  async createExpense(
+    expense: Omit<Expense, "id">,
+    uid: string
+  ): Promise<string> {
+    const collectionPath = getExpensesPath(expense.sharedExpenseId, uid);
+    const collectionRef = await getExpensesRef(collectionPath);
 
     let docRef;
 
@@ -91,16 +116,20 @@ export const expenseService = {
         createdAt: Timestamp.now(),
       });
 
-      const expenseList = await this.getExpenses(expense.sharedExpenseId);
+      const expenseList = await this.getExpenses(expense.sharedExpenseId, uid);
 
       let seTotalAmount = expenseList.reduce(
         (total, expense) => total + expense.amount,
         0
       );
 
-      await sharedExpenseService.update(expense.sharedExpenseId, {
-        totalAmount: seTotalAmount,
-      });
+      await sharedExpenseService.update(
+        expense.sharedExpenseId,
+        {
+          totalAmount: seTotalAmount,
+        },
+        uid
+      );
     });
 
     console.log("Transaction successfully committed! ==> ");
@@ -108,13 +137,13 @@ export const expenseService = {
   },
 
   // Get all expenses
-  async getExpenses(sharedExpenseId: string): Promise<Expense[]> {
+  async getExpenses(sharedExpenseId: string, uid: string): Promise<Expense[]> {
     if (sharedExpenseId === "") {
       return Promise.resolve([]);
     }
 
-    const collectionPath = getExpensesCollectionPath(sharedExpenseId);
-    const collectionRef = await getExpensesCollectionRef(collectionPath);
+    const collectionPath = getExpensesPath(sharedExpenseId, uid);
+    const collectionRef = await getExpensesRef(collectionPath);
 
     const querySnapshot = await getDocs(
       query(collectionRef, orderBy("date", "desc"))
@@ -130,9 +159,13 @@ export const expenseService = {
   },
 
   // Get expenses by participant
-  async getExpensesByUser(participantId: string): Promise<Expense[]> {
+  async getExpensesByUser(
+    participantId: string,
+    sharedExpenseId: string,
+    uid: string
+  ): Promise<Expense[]> {
     const q = query(
-      collection(db, EXPENSES_COLLECTION_NAME),
+      collection(db, getExpensesPath(sharedExpenseId, uid)),
       where("payerId", "==", participantId),
       orderBy("date", "desc")
     );
@@ -150,26 +183,29 @@ export const expenseService = {
   // Update expense
   async updateExpense(
     id: string,
-    updates: Partial<Omit<Expense, "id">>
+    updates: Partial<Omit<Expense, "id">>,
+    sharedExpenseId: string,
+    uid: string
   ): Promise<void> {
-    const docRef = doc(db, EXPENSES_COLLECTION_NAME, id);
+    const docRef = doc(db, getExpensesPath(sharedExpenseId, uid), id);
     await updateDoc(docRef, updates);
   },
 
   // Delete expense
   async deleteExpense(
     id: string,
-    currentSharedExpenseId: string
+    currentSharedExpenseId: string,
+    uid: string
   ): Promise<void> {
     if (currentSharedExpenseId === "") {
       return Promise.reject("No Current Shared Expense Selected");
     }
 
-    const docRef = doc(
-      db,
-      getExpensesCollectionPath(currentSharedExpenseId),
-      id
-    );
+    if (uid === "") {
+      return Promise.reject("No Current User Selected");
+    }
+
+    const docRef = doc(db, getExpensesPath(currentSharedExpenseId, uid), id);
     await deleteDoc(docRef);
   },
 };
@@ -178,38 +214,38 @@ export const expenseService = {
 
 // Payment Operations
 
-const PAYMENTS_COLLECTION_NAME = "payments";
-
-function getPaymentsCollectionPath(sharedExpenseId: string) {
-  return `environments/${
-    import.meta.env.VITE_FIRESTORE_DATA_ID
-  }/${SHARED_EXPENSES_COLLECTION_NAME}/${sharedExpenseId}/${PAYMENTS_COLLECTION_NAME}`;
+function getPaymentsPath(sharedExpenseId: string, uid: string) {
+  return `${BASE_COLLECTION_PATH}/${uid}/${SHARED_EXPENSES_COLLECTION_NAME}/${sharedExpenseId}/${PAYMENTS_COLLECTION_NAME}`;
 }
 
-async function getPaymentsCollectionRef(collectionPath: string) {
+async function getPaymentsRef(collectionPath: string) {
   return collection(db, collectionPath);
 }
 
 export const paymentService = {
-  async createPayment(payment: Omit<Payment, "id">): Promise<string> {
-    const paymentsCollectionRef = await getPaymentsCollectionRef(
-      getPaymentsCollectionPath(payment.sharedExpenseId)
+  async createPayment(
+    payment: Omit<Payment, "id">,
+    uid: string
+  ): Promise<string> {
+    const paymentsCollectionRef = await getPaymentsRef(
+      getPaymentsPath(payment.sharedExpenseId, uid)
     );
 
     const docRef = await addDoc(paymentsCollectionRef, {
       ...payment,
       createdAt: Timestamp.now(),
     });
+
     return docRef.id;
   },
 
-  async getPayments(sharedExpenseId: string): Promise<Payment[]> {
+  async getPayments(sharedExpenseId: string, uid: string): Promise<Payment[]> {
     if (sharedExpenseId === "") {
       return Promise.resolve([]);
     }
 
-    const paymentsCollectionRef = await getPaymentsCollectionRef(
-      getPaymentsCollectionPath(sharedExpenseId)
+    const paymentsCollectionRef = await getPaymentsRef(
+      getPaymentsPath(sharedExpenseId, uid)
     );
 
     const querySnapshot = await getDocs(
@@ -227,10 +263,12 @@ export const paymentService = {
 
   async getPaymentsBetweenUsers(
     user1Id: string,
-    user2Id: string
+    user2Id: string,
+    sharedExpenseId: string,
+    uid: string
   ): Promise<Payment[]> {
     const q = query(
-      collection(db, "payments"),
+      collection(db, getPaymentsPath(sharedExpenseId, uid)),
       where("fromId", "in", [user1Id, user2Id]),
       where("toId", "in", [user1Id, user2Id])
     );
@@ -247,17 +285,14 @@ export const paymentService = {
 
   async deletePayment(
     id: string,
-    currentSharedExpenseId: string
+    currentSharedExpenseId: string,
+    uid: string
   ): Promise<void> {
     if (currentSharedExpenseId === "") {
       return Promise.reject("No Current Shared Expense Selected");
     }
 
-    const docRef = doc(
-      db,
-      getPaymentsCollectionPath(currentSharedExpenseId),
-      id
-    );
+    const docRef = doc(db, getPaymentsPath(currentSharedExpenseId, uid), id);
     await deleteDoc(docRef);
   },
 };
@@ -266,23 +301,26 @@ export const paymentService = {
 
 // Shared Expenses Operation
 
-const sharedExpensesCollectionPath = `environments/${
-  import.meta.env.VITE_FIRESTORE_DATA_ID
-}/${SHARED_EXPENSES_COLLECTION_NAME}`;
+function getSharedExpensesPath(uid: string) {
+  return `${BASE_COLLECTION_PATH}/${uid}/${SHARED_EXPENSES_COLLECTION_NAME}`;
+}
 
-async function getSharedExpensesCollectionRef() {
-  return collection(db, sharedExpensesCollectionPath);
+function getSharedExpensesRef(collectionPath: string) {
+  return collection(db, collectionPath);
 }
 
 export const sharedExpenseService = {
-  create: async (data: Omit<SharedExpense, "id">): Promise<string> => {
-    const collectionRef = await getSharedExpensesCollectionRef();
+  create: async (
+    data: Omit<SharedExpense, "id">,
+    uid: string
+  ): Promise<string> => {
+    const collectionRef = getSharedExpensesRef(getSharedExpensesPath(uid));
     const docRef = await addDoc(collectionRef, data);
     return docRef.id;
   },
 
-  getAll: async (): Promise<SharedExpense[]> => {
-    const collectionRef = await getSharedExpensesCollectionRef();
+  getAll: async (uid: string): Promise<SharedExpense[]> => {
+    const collectionRef = getSharedExpensesRef(getSharedExpensesPath(uid));
     const snapshot = await getDocs(collectionRef);
     return snapshot.docs.map(
       (doc) => ({ id: doc.id, ...doc.data() } as SharedExpense)
@@ -291,9 +329,10 @@ export const sharedExpenseService = {
 
   update: async (
     id: string,
-    updates: Partial<SharedExpense>
+    updates: Partial<SharedExpense>,
+    uid: string
   ): Promise<void> => {
-    await updateDoc(doc(db, sharedExpensesCollectionPath, id), updates);
+    await updateDoc(doc(db, getSharedExpensesPath(uid), id), updates);
   },
 };
